bw_vs_pans/src/requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0

# AWS/SageMaker
boto3>=1.26.0
sagemaker>=2.130.0

# Data handling
pandas>=1.5.0

# Image processing
opencv-python>=4.8.0

# Metrics and evaluation
torchmetrics>=0.11.0

# Progress tracking
tqdm>=4.64.0

# Configuration management
PyYAML>=6.0

# Logging and monitoring
wandb>=0.15.0

# Visualization (for analysis)
matplotlib>=3.6.0
seaborn>=0.12.0

# For balanced sampling and statistics
scipy>=1.9.0